+++ src/VerticalTabsItem/docs/index.story.js
@@ -14,11 +14,10 @@ import {
   testkit,
 } from 'wix-storybook-utils/Sections';
 
-import icons from '../../../stories/utils/icons-for-story';
-import { storySettings } from './storySettings';
-import allComponents from '../../../stories/utils/allComponents';
-
 import VerticalTabsItem from '..';
+import { storySettings } from './storySettings';
+import icons from '../../../stories/utils/icons-for-story';
+import allComponents from '../../../stories/utils/allComponents';
 
 const code = config => baseCode({ components: allComponents, ...config });
 
@@ -30,9 +29,7 @@ export default {
   componentPath: '..',
 
   componentProps: {
-    onClick: () => {},
     children: 'Playground',
-    size: 'medium',
     type: 'tab',
   },
 
@@ -118,7 +115,7 @@ export default {
           columns([
             description({
               title: 'Vertical Tab With Suffix',
-              text: 'Vertical tab item with prefix icon',
+              text: 'Vertical tab item with suffix icon',
             }),
 
             code({