+++ CHANGELOG.md
@@ -18,10 +18,13 @@ Types of changes:
 
 ### Added
 - `<Checkbox/>` - add small size [#4525](https://github.com/wix/wix-style-react/pull/4525)
+- `<AutoCompleteWithLabel/>` - add `native` prop [#4528](https://github.com/wix/wix-style-react/pull/4528)
 
 ### Fixed
 - `<Table/>` - Fixed missing separator border for Table.Content when column titles are hidden [#4517](https://github.com/wix/wix-style-react/pull/4517)
 - `<BadgeSelect/>` - fix missing colors for `neutralLigth`, `warningLight` & `premium` skins [#4516](https://github.com/wix/wix-style-react/pull/4516)
+- `<PopoverMenu/>` (beta) - Fix zIndex prop not working as expected [#4530](https://github.com/wix/wix-style-react/pull/4530)
+- `<ImageViewer/>` - make sure that onImageLoad is called after the state has been changed [#4537](https://github.com/wix/wix-style-react/pull/4537)
 
 ## 7.15.1 - 2019 - 11 - 13